# openapi.yml
openapi: 3.0.0
paths:
  /api/users:
    get:
      operationId: UsersController_getAllUsers
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - users
  /api/users/me:
    get:
      operationId: UsersController_getMe
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - users
    patch:
      operationId: UsersController_updateUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - users
    delete:
      operationId: UsersController_deleteUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - users
  /api/auth/signup:
    post:
      operationId: AuthController_userSignup
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
      tags:
        - auth
  /api/auth/admin/signup:
    post:
      operationId: AuthController_adminSignup
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
      tags:
        - auth
  /api/auth/signin:
    post:
      operationId: AuthController_userLogin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninDto'
      responses:
        '201':
          description: ''
      tags:
        - auth
  /api/auth/admin/signin:
    post:
      operationId: AuthController_adminLogin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninDto'
      responses:
        '201':
          description: ''
      tags:
        - auth
  /api/auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - auth
  /api/categories:
    post:
      operationId: CategoriesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Category successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '401':
          description: Unauthorized – missing or invalid token
        '403':
          description: Forbidden – insufficient role
        '409':
          description: Category with the same name already exists
      security:
        - bearer: []
      summary: Create a new category
      tags:
        - categories
    get:
      operationId: CategoriesController_findAll
      parameters: []
      responses:
        '200':
          description: List of all categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponseDto'
        '401':
          description: Unauthorized – missing or invalid token
        '403':
          description: Forbidden – insufficient role
      security:
        - bearer: []
      summary: Get all categories
      tags:
        - categories
  /api/categories/{id}:
    get:
      operationId: CategoriesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Category ID
          schema:
            type: number
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '401':
          description: Unauthorized – missing or invalid token
        '403':
          description: Forbidden – insufficient role
        '404':
          description: Category not found
      security:
        - bearer: []
      summary: Get a category by ID
      tags:
        - categories
    patch:
      operationId: CategoriesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Category ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Category successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '401':
          description: Unauthorized – missing or invalid token
        '403':
          description: Forbidden – insufficient role
        '404':
          description: Category not found
      security:
        - bearer: []
      summary: Update a category by ID
      tags:
        - categories
    delete:
      operationId: CategoriesController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Category ID
          schema:
            type: number
      responses:
        '200':
          description: Category successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '401':
          description: Unauthorized – missing or invalid token
        '403':
          description: Forbidden – insufficient role
        '404':
          description: Category not found
      security:
        - bearer: []
      summary: Delete a category by ID
      tags:
        - categories
info:
  title: Ai nutrition API
  description: API for managing ai nutrition API
  version: '1.0'
  contact: {}
tags:
  - name: categories
    description: ''
  - name: users
    description: ''
  - name: meals
    description: ''
  - name: ai
    description: ''
  - name: auth
    description: ''
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    UpdateUserDto:
      type: object
      properties: {}
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
          example: John Doe
          description: Full name of the user
        email:
          type: string
          example: test@example.com
          description: User email address
        password:
          type: string
          example: strongPassword123
          description: User password
        age:
          type: number
          example: 25
          description: User age in years
        weight:
          type: number
          example: 70.5
          description: User weight in kg
        height:
          type: number
          example: 175
          description: User height in cm
        goal:
          type: string
          example: lose fat
          description: User fitness goal
        preferences:
          type: string
          example: vegetarian, non-dairy
          description: User dietary preferences
      required:
        - name
        - email
        - password
        - age
        - weight
        - goal
    SigninDto:
      type: object
      properties:
        user_id:
          type: number
          example: 12
        password:
          type: string
          example: strongPassword123
      required:
        - user_id
        - password
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: string
          example: Fruits
          description: Unique name of the category (e.g., Fruits, Vegetables, Meats)
        description:
          type: string
          example: Plant-based foods that are sweet and rich in simple carbohydrates.
          description: >-
            A short description explaining what type of foods belong to this
            category.
      required:
        - name
    CategoryResponseDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: frutas
        description:
          type: string
          example: Plant-based foods that are sweet and rich in simple carbohydrates.
          nullable: true
        created_at:
          format: date-time
          type: string
          example: '2025-11-09T09:00:00.000Z'
          description: Date when the category was created
        updated_at:
          format: date-time
          type: string
          example: '2025-11-09T09:30:00.000Z'
          description: Date when the category was last updated
      required:
        - id
        - name
        - description
        - created_at
        - updated_at
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: string
          example: Fruits
          description: Updated category name (optional, must remain unique if changed).
        description:
          type: string
          example: Plant-based foods that are sweet and rich in simple carbohydrates.
          description: Updated description for the category (optional).